Only invoke didTouchMemory() when a memory write occurred.
This used to be done for every memoryPointer() retrieval, but since
those have been refactored to return const pointers, we can now allow
ourselves to assume that all memory writes are routed through
CPU::writePhysicalMemory().

I'd like a better design for this somehow but I'm not sure what to do.