Move hardware devices to the worker thread.
This probably has some bugs, but it was rather broken the way it was.
The idea here is that all hardware will live on the worker thread,
and any communication between them and the GUI will have to be done
through some kinda command/request queue.

This patch introduces a ThreadedTimer helper thingy for the PIT which
I will use for more things soon.

I also generalized the IODevicePass thingy into a Badge template.
I find this pretty cute: it allows you to define a public function
that takes a Badge<Foo>, effectively making it so that only member
functions of Foo are able to call that function.